Add half-star ratings and StarRating UI
Switch score handling from integer (1–10) to half-star floats (0.5–5.0) and add a StarRating component. Backend: change SQLAlchemy column to Float and update Pydantic schemas/validator to accept half-star increments; API now expects float scores in 0.5 steps. Frontend: introduce StarRating, update BookForm to use numeric/null score state with interactive half-star selection and stricter validation, update BookCard to display the star UI and adjust pages badge styling. Note: BookPayload.score was changed to a required number and the DB column type changed — run a DB migration or adjust data as needed to avoid breakage.

+++ backend/models.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 from typing import Optional
-from sqlalchemy import String, Text, Integer, DateTime
+from sqlalchemy import String, Text, Integer, Float, DateTime
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column
 from database import Base
@@ -14,7 +14,7 @@ class Book(Base):
     title: Mapped[str] = mapped_column(String, nullable=False)
     author: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     review: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    score: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    score: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     cover_url: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     pages: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
